Si7005: Fixed a couple of bugs regarding initialisation.

+++ src/si7005/si7005.h
@@ -152,7 +152,7 @@ class SI7005 : public ITemperatureHumiditySensor {
         int m_pin;
         mraa_i2c_context m_i2cControlCtx;
         bool configured;
-        int config_reg;
+        uint8_t config_reg;
         float last_temperature;
 
         int getMeasurement(uint8_t configValue);
